Delete non-existing records logic.

+++ resource-service/src/main/java/com/epam/learn/resource/service/impl/ResourceServiceImpl.java
@@ -60,10 +60,13 @@ public class ResourceServiceImpl implements ResourceService {
     public DeleteResourcesResponse deleteResources(String idsCsv) {
         List<Long> requestIds = extractIds(idsCsv);
         Iterable<ResourceEntity> existingEntities = resourceRepository.findAllById(requestIds);
-        List<Long> existingIds = extractIds(existingEntities);
 
-        songsApi.deleteSongs(joinIds(existingIds));
-        resourceRepository.deleteAllById(existingIds);
+        if (existingEntities.iterator().hasNext()) {
+            List<Long> existingIds = extractIds(existingEntities);
+
+            songsApi.deleteSongs(joinIds(existingIds));
+            resourceRepository.deleteAllById(existingIds);
+        }
 
         return resourceMapper.toDeleteResourceResponse(existingEntities);
     }
